Consider Delay in Replication
If a master has not replicated at all or not for a long time and a slave
initiates the warmup, both offsets are for short period of time 0 until
the first data is transfered to the slave. In order to not stop replication
to early, don't accept 0 vales for the master offset.